separated verifiable process init into new file.
